chore: rename TanStackTable.tsx to GenericTable.tsx
